+++ specs/phase-2/backend/app/db.py
@@ -1,0 +1,10 @@
+from sqlmodel import create_engine, SQLModel  
+from .models.user import User  
+from .models.todo import Todo  
+  
+DATABASE_URL = 'sqlite:///./todo_app.db'  
+  
+engine = create_engine(DATABASE_URL)  
+  
+def create_db_and_tables():  
+    SQLModel.metadata.create_all(engine)