+++ specs/phase-2/backend/migrations/versions/e2e4f020bb8a_initial_models.py
@@ -1,0 +1,28 @@
+"""Initial models
+
+Revision ID: e2e4f020bb8a
+Revises: 
+Create Date: 2026-01-09 19:17:37.061192
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'e2e4f020bb8a'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    pass
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    pass